# conf/swagger.yaml
swagger: '2.0'
info:
  title: 'Node.js course : Phase 3 : Swagger'
  description: A bit more complicated sample project
  version: '0.3.0'
host: localhost:3000
schemes:
  - http
basePath: /
paths:
...
  '/books/{bookId}':
    get:
      description: Returns book info, including its authors, based on a single ID
      tags:
        - books
      produces:
        application/json
      responses:
        '200':
          description: book information
      parameters:
        - name: bookId
          in: path
          description: ID of the book to fetch
          required: true
          type: integer
  '/books/{bookId}/reviews':
    get:
      description: Returns posted user reviews, based on a single book ID
      tags:
        - books
      produces:
        application/json
      responses:
        '200':
          description: user reviews
      parameters:
        - name: bookId
          in: path
          description: ID of the book to fetch user reviews to
          required: true
          type: integer
...